- Accept CHANGELOG.md deletion (a6e6fd39); task-33 behavior docs live on
  in docsrc/manual/shortcuts.md and piano-roll.md.
- Reconcile absent-loop sentinels to the Tick world: MidiTimeline
  Tick-sourced comparisons use CoreTimeDefaults::kNoTick; TimeAxis
  uint64 accessors map kNoTick to UINT64_MAX at the boundary.
- Event editor rejects tick input above kMaxTick at model and
  controller layers; high-bit tests prove rejection + kMaxTick exactness.
- Clip rescale tempo arm saturates at kMaxTick (Tick-typed output).

+++ src/checks/editcheck/tst_songdocument_songtime.cpp
@@ -76,7 +76,7 @@ void EditCheckTest::songWholeSongRemove()
     const auto maxEnd = [&document] {
         uint64_t end = 0;
         for (const SmfTrack &chunk : document.smf().tracks)
-            end = std::max(end, chunk.endTick);
+            end = std::max(end, uint64_t(chunk.endTick));
         return end;
     };
     const TempoPoint originalTempo = songdocument_test::tempo(base + step * 63, 150);
